Move VoidableType to own file
